About Me
Added content to about me page. Edited about.html to write text about me. Information on the page still needs to be fleshed out and I will add more the next coming days. I want to get this on my github page so I can push it to the web